+++ src/Microsoft.ApplicationInsights/Extensibility/Implementation/OperationTelemetry.cs
@@ -74,6 +74,11 @@
         /// </summary>
         public abstract string Sequence { get; set; }
 
+        /// <summary>
+        /// Gets or sets gets the extension used to extend this telemetry instance using new strong typed object.
+        /// </summary>
+        public abstract IExtension Extension { get; set; }
+
         /// <summary>  
         /// Gets or sets Time in StopWatch ticks representing begin time of the operation. Used internally
         /// for calculating duration between begin and end.
@@ -93,6 +98,9 @@
         /// </summary>
         /// <returns>A cloned instance.</returns>
         public abstract ITelemetry DeepClone();
+
+        /// <inheritdoc/>
+        public abstract void SerializeData(ISerializationWriter serializationWriter);
 
         /// <summary>
         /// Sets operation Id.